tests/wpt/tests/css/css-shapes/shape-functions/rect-function-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Shapes Module Level 1: parsing the rect() function</title>
<link rel="help" href="https://drafts.csswg.org/css-shapes-1/#funcdef-basic-shape-rect">
<meta name="assert" content="Tests parsing of the inset() function">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("clip-path", "rect()");
test_invalid_value("clip-path", "rect(0px)");
test_invalid_value("clip-path", "rect(0px 1px)");
test_invalid_value("clip-path", "rect(0px 1px 2px)");
test_invalid_value("clip-path", "rect(0px, 1px, 2px, 3px)");
test_invalid_value("clip-path", "rect(0, 1, 2, 3)");
test_invalid_value("clip-path", "rect(1% 2% 3% 4% 5%)");
test_invalid_value("clip-path", "rect(round 0)");
test_invalid_value("clip-path", "rect(10px auto 20px 30px round)");
test_invalid_value("clip-path", "rect(10px auto 20px 30px round 123)");
test_invalid_value("clip-path", "rect(10px auto 20px 30px round 1% 2% 3% 4% 5%)");
test_invalid_value("clip-path", "rect(10px auto 20px 30px round / 1px)");
test_invalid_value("clip-path", "rect(10px auto 20px 30px round -20px)");
test_invalid_value("clip-path", "rect(10px auto 20px 30px round -40%)");
</script>
</body>
</html>
